Revise measures list, add reduced results option
The measures list was revised to reflect the latest for the decarbonization pathways project.

Additionally, users may write out a more concise version of results in run.py (focused on energy, carbon, and cost savings, potentially for focus years) via the --trim_results command line option.

A minor adjustment to the packaging in ecm_prep.py constrains the envelope relative performance multiplier between the 0 and 1 range.